# internal/rest/openapi3.json
{"components":{"requestBodies":{"CreateTasksRequest":{"content":{"application/json":{"schema":{"properties":{"dates":{"$ref":"#/components/schemas/Dates"},"description":{"minLength":1,"type":"string"},"priority":{"$ref":"#/components/schemas/Priority"}}}}},"description":"request to create a new task.","required":true},"UpdateTasksRequest":{"content":{"application/json":{"schema":{"properties":{"dates":{"$ref":"#/components/schemas/Dates"},"description":{"minLength":1,"type":"string"},"is_done":{"type":"boolean"},"priority":{"$ref":"#/components/schemas/Priority"}}}}},"description":"request to update an existing task.","required":true}},"responses":{"CreateTasksResponse":{"content":{"application/json":{"schema":{"properties":{"task":{"$ref":"#/components/schemas/Task"}}}}},"description":"Response returned back after creating tasks."},"ErrorResponse":{"content":{"application/json":{"schema":{"properties":{"error":{"type":"string"},"status":{"type":"integer"}}}}},"description":"Response when errors happens"},"ReadTasksResponse":{"content":{"application/json":{"schema":{"properties":{"task":{"$ref":"#/components/schemas/Task"}}}}},"description":"Response returned back after searching one task."}},"schemas":{"Dates":{"properties":{"due":{"format":"date-time","nullable":true,"type":"string"},"start":{"format":"date-time","nullable":true,"type":"string"}},"type":"object"},"Priority":{"default":"none","enum":["none","low","medium","high"],"type":"string"},"Task":{"properties":{"dates":{"$ref":"#/components/schemas/Dates"},"description":{"type":"string"},"id":{"format":"uuid","type":"string"},"is_done":{"type":"boolean"},"priority":{"$ref":"#/components/schemas/Priority"}},"type":"object"}}},"info":{"contact":{"url":"https://github.com/Akshit8/tdm"},"description":"REST API to create and manage TODO's","license":{"name":"MIT","url":"https://opensource.org/licenses/MIT"},"title":"Todo API","version":"0.0.1"},"openapi":"3.0.0","paths":{"/tasks":{"post":{"operationId":"CreateTask","requestBody":{"$ref":"#/components/requestBodies/CreateTasksRequest"},"responses":{"201":{"$ref":"#/components/responses/CreateTasksResponse"},"400":{"$ref":"#/components/responses/ErrorResponse"},"500":{"$ref":"#/components/responses/ErrorResponse"}}}},"/tasks/{taskId}":{"delete":{"operationId":"DeleteTask","parameters":[{"in":"path","name":"taskId","required":true,"schema":{"format":"uuid","type":"string"}}],"responses":{"200":{"description":"Response when errors happens"},"404":{"$ref":"#/components/responses/ErrorResponse"},"500":{"$ref":"#/components/responses/ErrorResponse"}}},"get":{"operationId":"ReadTask","parameters":[{"in":"path","name":"taskId","required":true,"schema":{"format":"uuid","type":"string"}}],"responses":{"200":{"$ref":"#/components/responses/ReadTasksResponse"},"404":{"$ref":"#/components/responses/ErrorResponse"},"500":{"$ref":"#/components/responses/ErrorResponse"}}},"put":{"operationId":"UpdateTask","parameters":[{"in":"path","name":"taskId","required":true,"schema":{"format":"uuid","type":"string"}}],"requestBody":{"$ref":"#/components/requestBodies/UpdateTasksRequest"},"responses":{"200":{"description":"Task updated"},"400":{"$ref":"#/components/responses/ErrorResponse"},"404":{"$ref":"#/components/responses/ErrorResponse"},"500":{"$ref":"#/components/responses/ErrorResponse"}}}}},"servers":[{"description":"Local development","url":"http://127.0.0.1:8000"}]}